VoltDB Configuration as a Spring singleton
