+++ src/frontend/org/voltdb/Inits.java
@@ -46,6 +46,7 @@ import org.voltcore.utils.Pair;
 import org.voltdb.catalog.Catalog;
 import org.voltdb.common.Constants;
 import org.voltdb.compiler.deploymentfile.DeploymentType;
+import org.voltdb.config.Configuration;
 import org.voltdb.export.ExportManager;
 import org.voltdb.iv2.MpInitiator;
 import org.voltdb.iv2.TxnEgo;
@@ -68,7 +69,7 @@ public class Inits {
     private static final VoltLogger hostLog = new VoltLogger("HOST");
 
     final RealVoltDB m_rvdb;
-    final VoltDB.Configuration m_config;
+    final Configuration m_config;
     final boolean m_isRejoin;
     DeploymentType m_deployment = null;
